// frontend/src/core/api/generated/model/siteEventView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { SiteEventViewKind } from './siteEventViewKind'

export interface SiteEventView {
  id: string
  siteId: string
  kind: SiteEventViewKind
  /** @nullable */
  actorId?: string | null
  /** @nullable */
  actorName?: string | null
  /** @nullable */
  subjectId?: string | null
  /** @nullable */
  subjectName?: string | null
  createdAt: string
}